Split up these classes to make it easier to simplify
